reverted some unwanted commited changes
