--- lib/KDropdownMenu/__tests__/KDropdownMenu.spec.js
@@ -1,29 +1,0 @@
-// eslint-disable-next-line import/named
-import { renderComponent, takeSnapshot, click } from '../../../jest.conf/visual.testUtils';
-
-describe.visual('KDropdownMenu Visual Tests', () => {
-  const snapshotOptions = {
-    widths: [400],
-    minHeight: 512,
-  };
-
-  it('renders KDropdownMenu variants', async () => {
-    await renderComponent('KDropdownMenuTest');
-
-    // Click the first button to open its dropdown
-    await click('.test-case:nth-child(1) button');
-    await takeSnapshot('KDropdownMenu - Single Item', snapshotOptions);
-
-    // Click the second button to open its dropdown
-    await click('.test-case:nth-child(2) button');
-    await takeSnapshot('KDropdownMenu - Multiple Items', snapshotOptions);
-
-    // Click the third button to open its dropdown
-    await click('.test-case:nth-child(3) button');
-    await takeSnapshot('KDropdownMenu - With Icons', snapshotOptions);
-
-    // Click the fourth button to open its dropdown
-    await click('.test-case:nth-child(4) button');
-    await takeSnapshot('KDropdownMenu - With Header', snapshotOptions);
-  });
-});
